+++ src/react/src/components/InitialClaimFlow/ClaimForm/ClaimForm.jsx
@@ -18,7 +18,7 @@ import Business from '@material-ui/icons/Business';
 import ClaimFormStepper from './Stepper/Stepper';
 import EligibilityStep from './Steps/EligibilityStep/EligibilityStep';
 import ContactStep from './Steps/ContactStep';
-import BusinessStep from './Steps/BusinessStep';
+import BusinessStep from './Steps/BusinessStep/BusinessStep';
 import ProfileStep from './Steps/ProfileStep/ProfileStep';
 import ErrorState from './ErrorState/ErrorState';
 import RequireAuthNotice from '../../RequireAuthNotice';
@@ -154,6 +154,7 @@ const ClaimForm = ({
         claimForm,
         handleFieldChange,
         handleBlur,
+        updateFieldWithoutTouch,
         isButtonDisabled,
     } = useClaimForm(
         formData,
@@ -289,6 +290,7 @@ const ClaimForm = ({
                             formData={claimForm.values}
                             handleChange={handleFieldChange}
                             handleBlur={handleBlur}
+                            updateFieldWithoutTouch={updateFieldWithoutTouch}
                             errors={claimForm.errors}
                             touched={claimForm.touched}
                             countryOptions={countriesOptions}